all/requirements.txt
albumentations==0.4.5
appdirs==1.4.3
atpublic==1.0
ax-platform==0.1.9
botorch==0.2.1
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
cloudpickle==1.4.1
colorama==0.4.3
configobj==5.0.6
coremltools==3.3
cycler==0.10.0
decorator==4.4.2
distro==1.5.0
dpath==2.0.1
dvc==0.93.0
flatten-json==0.1.7
flufl.lock==3.2
funcy==1.14
future==0.18.2
gitdb==4.0.4
GitPython==3.1.1
gpytorch==1.1.1
grandalf==0.6
humanize==2.4.0
hyperopt==0.2.4
idna==2.9
imageio==2.8.0
imgaug==0.2.6
importlib-metadata==1.6.0
inflect==3.0.2
Jinja2==2.11.2
joblib==0.14.1
jsonpath-ng==1.5.1
kiwisolver==1.2.0
MarkupSafe==1.1.1
matplotlib==3.2.1
mpmath==1.1.0
nanotime==0.5.2
networkx==2.3
numpy==1.18.3
onnx==1.6.0
onnx-coreml==1.3
opencv-python-headless==4.2.0.34
packaging==20.3
pandas==1.0.3
pathspec==0.8.0
Pillow==7.1.2
plotly==4.6.0
ply==3.11
protobuf==3.11.3
pyasn1==0.4.8
pydot==1.4.1
pygtrie==2.3.2
pyparsing==2.4.7
python-dateutil==2.8.0
pytz==2020.1
PyWavelets==1.1.1
PyYAML==5.3.1
requests==2.23.0
retrying==1.3.3
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
scikit-image==0.16.2
scikit-learn==0.22.2.post1
scipy==1.4.1
shortuuid==1.0.1
six==1.14.0
smmap==3.0.2
sympy==1.5.1
texttable==1.6.2
torch==1.5.0
torchvision==0.6.0
tqdm==4.45.0
treelib==1.6.1
typing==3.7.4.1
typing-extensions==3.7.4.2
urllib3==1.25.9
voluptuous==0.11.7
zc.lockfile==2.0
zipp==3.1.0
